Complete Skeleton for digging and dumping
The skeleton code for the digging and dumping operations are complete.

The skeleton code for the locomotion operation still needs to be finished.

Questionable implementation for stopping both the depth controls and the auger controls from last year's code under the digging.py file. Future testing and discussing required on problem to determine possible simplification of code.