main files completed
This commit includes the main.py, main.launch, and main.wrapper.py files. The wrapper file has been edited to accommodate this years robot in the GUI. Some there are a few GUI design features that will remain in the code for the time being until the robot can be tested with the GUI changes.