catkin changes
Includes the changes to teleop in order to have press and hold controls for manual control of robot.

Along with the Encoder_test.py file, that was made for verification testing of the motors and the accuracy of the encoders for distance traveled.

The large amount of files that are included in this push are a result of having to re-make the catkin workspace due to an error that Abraham and Chris ran into.